+++ Marlin/src/gcode/feature/mixing/M165.cpp
@@ -20,19 +20,26 @@
  *
  */
 
+#include "../../../inc/MarlinConfig.h"
+
+#if ENABLED(DIRECT_MIXING_IN_G1)
+
+#include "../../gcode.h"
+#include "../../../feature/mixing.h"
+
 /**
- * M163: Set a single mix factor for a mixing extruder
- *       This is called "weight" by some systems.
+ * M165: Set multiple mix factors for a mixing extruder.
+ *       Factors that are left out will be set to 0.
+ *       All factors together must add up to 1.0.
  *
- *   S[index]   The channel index to set
- *   P[float]   The mix value
+ *   A[factor] Mix factor for extruder stepper 1
+ *   B[factor] Mix factor for extruder stepper 2
+ *   C[factor] Mix factor for extruder stepper 3
+ *   D[factor] Mix factor for extruder stepper 4
+ *   H[factor] Mix factor for extruder stepper 5
+ *   I[factor] Mix factor for extruder stepper 6
  *
  */
-void gcode_M163() {
-  const int mix_index = parser.intval('S');
-  if (mix_index < MIXING_STEPPERS) {
-    float mix_value = parser.floatval('P');
-    NOLESS(mix_value, 0.0);
-    mixing_factor[mix_index] = RECIPROCAL(mix_value);
-  }
-}
+void GcodeSuite::M165() { gcode_get_mix(); }
+
+#endif // DIRECT_MIXING_IN_G1
